chore: rename logging
